requirements.txt
async-timeout==4.0.2
cachelib==0.9.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Caching==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
redis==4.4.0
requests==2.28.1
urllib3==1.26.13
Werkzeug==2.2.2
